Refactor: ChattingErrorCode 위치 errorcode 패키지로 바꾸기
